Card details component from Fragment added
