+++ ni_sdk/src/main/java/ae/network/nicardmanagementsdk/presentation/ui/card_details/fragment/CardDetailsFragmentFromActivity.kt
@@ -1,0 +1,28 @@
+package ae.network.nicardmanagementsdk.presentation.ui.card_details.fragment
+
+import ae.network.nicardmanagementsdk.api.models.input.NIInput
+import ae.network.nicardmanagementsdk.presentation.models.Extra
+import android.content.Context
+import android.os.Bundle
+
+class CardDetailsFragmentFromActivity: CardDetailsFragmentBase() {
+
+    override fun checkSubscriber(context: Context) {
+        if (context is OnFragmentInteractionListener) {
+            listener = context
+        } else {
+            throw RuntimeException("${context::class.java.simpleName} must implement CardDetailsFragmentBase.OnFragmentInteractionListener")
+        }
+    }
+
+    companion object {
+        @JvmStatic
+        fun newInstance(input: NIInput) = CardDetailsFragmentFromActivity().apply {
+            arguments = Bundle().apply {
+                putSerializable(Extra.EXTRA_NI_INPUT, input)
+            }
+        }
+
+        const val TAG = "CardDetailsFragmentFromActivity"
+    }
+}
