+++ ni_sdk/src/main/res/layout/fragment_card_details.xml
@@ -37,7 +37,7 @@
             style="@style/Widget.NICardManagementSDK.TextView.Transparency"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="@string/card_number"
+            android:text="@string/card_number_label"
             android:textSize="8sp"
             android:textStyle="normal"
             app:layout_constraintBottom_toTopOf="@+id/guideline1_horizontal"
@@ -79,7 +79,7 @@
             style="@style/Widget.NICardManagementSDK.TextView.Transparency"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="@string/expiry_date"
+            android:text="@string/expiry_date_label"
             android:textSize="8sp"
             android:textStyle="normal"
             app:layout_constraintBottom_toTopOf="@id/guideline3_horizontal"
@@ -106,7 +106,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="40dp"
-            android:text="@string/cvv"
+            android:text="@string/cvv_label"
             android:textSize="8sp"
             android:textStyle="normal"
             app:layout_constraintBaseline_toBaselineOf="@id/expiryDateLabelTextView"
@@ -146,7 +146,7 @@
             style="@style/Widget.NICardManagementSDK.TextView.Transparency"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="@string/name"
+            android:text="@string/name_label"
             android:textSize="8sp"
             android:textStyle="normal"
             app:layout_constraintBottom_toTopOf="@id/guideline5_horizontal"
